\name{align_genomes}
\alias{align_genomes}
\title{Align genomes}
\usage{
  align_genomes(seg_dir, aligner = "fsa", force = FALSE,
    ...)
}
\arguments{
  \item{seg_dir}{Path to mercator segments.}

  \item{aligner}{Which alignment program to use; 'fsa' or
  'mavid'.}

  \item{...}{Other params.}
}
\description{
  Align genomes
}

