% Generated by roxygen2 (4.0.0): do not edit by hand
\name{alignment2Genome}
\alias{alignment2Genome}
\title{Map alignment positions to genomic positions.}
\usage{
alignment2Genome(ranges, aln, targetGenomes = NULL)
}
\arguments{
\item{ranges}{a \code{\linkS4class{GRanges}} or
\code{\linkS4class{GRangesList}} containing alignment
positions to map to the \code{targetGenomes}.}

\item{aln}{an \code{\linkS4class{annotatedAlignment}} instance.}

\item{targetGenomes}{a character vector of genome names.}
}
\value{
A \code{\linkS4class{GRangesList}} object.
}
\description{
Map alignment positions to genomic positions.
}
\seealso{
\code{\link{genome2Alignment}}, \code{\link{genome2Genome}},
\code{\link{sliceAlignment}}, \code{\link{findAnnotation}}.
}

