% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{class}
\name{annotatedAlignment-class}
\alias{annotatedAlignment-class}
\title{Construct an \sQuote{annotatedAlignment}.}
\description{
\dQuote{annotatedAlignment} is an S4 class that provides a container for
multiple alignments stored as an \dQuote{\linkS4class{XStringSet}} object
and genomic annotations stored as an \dQuote{\linkS4class{annotationList}}
object.
}
\section{Slots}{

\describe{
\item{\code{annotation}}{An \code{\linkS4class{annotationList}} object.}

\item{\code{alignment}}{An \code{\linkS4class{XStringSet}} object.}
}}

