% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{class}
\name{annotationList-class}
\alias{annotationList-class}
\title{Construct an \dQuote{annotationList}.}
\description{
\dQuote{annotationList} is an S4 class that provides a container for
genomic annotations stored in \dQuote{\linkS4class{GRanges}}.
}

