\name{fsa}
\alias{fsa}
\title{Sequence alignment with fsa}
\usage{
  fsa(seqfile, outfile = "fsa.mfa", ...)
}
\arguments{
  \item{seqfile}{Sequence files in fasta format.}

  \item{outfile}{Multifasta alignment file.}

  \item{...}{Additional arguments to fsa}
}
\description{
  Sequence alignment with fsa
}

