\docType{package}
\name{genoslideR}
\alias{genoslideR}
\alias{genoslideR-package}
\title{Genome Alignments and Annotation.}
\description{
  Handle genome alignments with annotation in R
}
\details{
  \tabular{ll}{ Package: \tab genoslideR\cr Type: \tab
  Package\cr Version: \tab 0.1-0\cr Date: \tab
  2012-11-05\cr License: \tab BSD\cr }
}
\author{
  Gerhard Schoefl \email{gschof@yahoo.de}, Jochen Bick
  \email{j.bick@uni-jena.de}

  Maintainer: Gerhard Schoefl \email{gschof@yahoo.de}
}
\keyword{package}

