\name{glimmer3}
\alias{glimmer3}
\title{Ab initio genome annotation using glimmer3}
\usage{
  glimmer3(genome,
    glimmeropts = list(o = 50, g = 110, t = 30),
    cleanup = TRUE)
}
\arguments{
  \item{genome}{Path to a genome file in fasta format}

  \item{glimmeropts}{Options for glimmer3}

  \item{cleanup}{Clean up intermediate files.}
}
\description{
  Ab initio genome annotation using glimmer3
}

