\name{importAlignment}
\alias{importAlignment}
\title{Import aligned mercator segments or a multifasta alignment}
\usage{
  importAlignment(alnPath, writeToMFA = TRUE)
}
\arguments{
  \item{alnPath}{Path to mercator segments or a multifasta
  segment}

  \item{writeToMFA}{if \code{TRUE} merge mercator segments
  into a single multi fasta file and write to file
  "aln.mfa"}
}
\description{
  Import aligned mercator segments or a multifasta
  alignment
}

