\name{map2aln}
\alias{map2aln}
\title{INTERNAL: Map a genomic range to the alignment}
\usage{
  map2aln(ranges, gmap, amap, gaps)
}
\arguments{
  \item{range}{A \code{\linkS4class{Granges}} object.}

  \item{gmap}{}

  \item{amap}{}

  \item{gaps}{}
}
\description{
  INTERNAL: Map a genomic range to the alignment
}
\keyword{internal}

