\name{mavid}
\alias{mavid}
\title{Sequence alignment with mavid}
\usage{
  mavid(seqfile, treefile = "treefile",
    outfile = "mavid.mfa", ...)
}
\arguments{
  \item{seqfile}{Sequence files in multifasta format.}

  \item{treefile}{Phylogenetic tree file in Newick format.}

  \item{outfile}{Multifasta alignment file.}

  \item{...}{Additional arguments to mavid}
}
\description{
  Sequence alignment with mavid
}

