% Generated by roxygen2 (4.0.0): do not edit by hand
\name{plot.OrthoMatrix}
\alias{plot.OrthoMatrix}
\title{Plot an orthology matrix}
\usage{
\method{plot}{OrthoMatrix}(x, i = NULL, ...)
}
\arguments{
\item{omat}{A list of orthology matrices.}

\item{i}{Index.}
}
\description{
Plot an orthology matrix
}

