% Generated by roxygen2 (4.0.0): do not edit by hand
\name{rbh}
\alias{rbh}
\title{Find Reciprocal Best Hits using Blat}
\usage{
rbh(seq_files, anno_files = NULL, anno_type = "glimmer3",
  blatopts = list(), glimmeropts = list(o = 50, g = 110, t = 30),
  wd = NULL)
}
\arguments{
\item{seq_files}{Path to sequence files ('fasta' or 'genbank').}

\item{anno_files}{(Optional) Path to annotation files. If no
annotation files are provided an \emph{ab initio} annotation
using \code{\link{glimmer3}} is performed.}

\item{anno_type}{Type of annotation ('glimmer3', 'genbank', 'gff',
'ptt', or 'ftable').}

\item{blatopts}{Options passed to \code{blat}.}

\item{glimmeropts}{Options passed to \code{\link{glimmer3}} if an
\emph{ab initio} annotation is performed.}

\item{wd}{Working directory. Defaults to the parent directory of
the provided sequence files.}
}
\description{
Find Reciprocal Best Hits using Blat
}
\details{
\code{rbh} will create a hidden directory \sQuote{.rbh}
in the parent directory of the submitted sequence files and place
all results of the intermediate computational steps in this directory.

The results of a \code{rbh} run are stored in a directory
\sQuote{rbh} in the parent directory of the submitted sequence files.
}

