\name{slice_aln}
\alias{slice_aln}
\title{INTERNAL: Slice genomic ranges mapped to an alignment}
\usage{
  slice_aln(ranges, aln, targetGenomes = NULL)
}
\arguments{
  \item{ranges}{A \code{\linkS4class{Granges}} object.}

  \item{aln}{A \code{\linkS4class{AnnotatedAlignment}}
  object.}
}
\description{
  INTERNAL: Slice genomic ranges mapped to an alignment
}
\keyword{internal}

