\name{sliceAlnranges}
\alias{sliceAlnranges}
\title{[INTERNAL] Slice genomic ranges mapped to an alignment}
\usage{
  sliceAlnranges(alnranges, aln)
}
\arguments{
  \item{ranges}{A \code{\linkS4class{Granges}} object.}

  \item{aln}{A \code{\linkS4class{AnnotatedAlignment}}
  object.}
}
\description{
  [INTERNAL] Slice genomic ranges mapped to an alignment
}
\keyword{internal}

