% Generated by roxygen2 (4.0.0): do not edit by hand
\name{sliceGRL}
\alias{sliceGRL}
\title{[INTERNAL] Slice GRangesList from an alignment}
\usage{
sliceGRL(grl, dss, compact = FALSE)
}
\arguments{
\item{grl}{A \code{\linkS4class{GrangesList}} object.}

\item{dss}{A \code{\linkS4class{DNAStringSet}} object.}
}
\description{
[INTERNAL] Slice GRangesList from an alignment
}
\keyword{internal}

