\name{substituteCDSNumbers}
\alias{substituteCDSNumbers}
\title{Replace CDS indices in an oframe by the corresponding gbFeatureList indeces.}
\usage{
  substituteCDSNumbers(oframe, flist)
}
\arguments{
  \item{oframe}{}

  \item{flist}{A (list of)
  \code{\link[biofiles]{gbFeatureList}}s}
}
\description{
  Replace CDS indices in an oframe by the corresponding
  gbFeatureList indeces.
}

