\name{alignMercatorSegments}
\alias{alignMercatorSegments}
\title{Align orthologous genome segments}
\usage{
  alignMercatorSegments(seg_dir, aligner = "fsa",
    force = FALSE, mask = TRUE)
}
\arguments{
  \item{seg_dir}{Path to mercator segments.}

  \item{aligner}{Which alignment program to use; 'fsa' or
  'mavid'.}

  \item{...}{Other params.}
}
\description{
  Align orthologous genome segments
}

