\name{fsa}
\alias{fsa}
\title{Sequence alignment with fsa}
\usage{
  fsa(seqfile, outfile = "fsa.mfa", opts = list(), ...)
}
\arguments{
  \item{seqfile}{Sequence files in fasta format.}

  \item{outfile}{Multifasta alignment file.}

  \item{opts}{a named list of options for fsa.}

  \item{...}{Named values interpreted as options for fsa.}
}
\description{
  Sequence alignment with fsa
}

