% Generated by roxygen2 (4.0.0): do not edit by hand
\name{genome2Genome}
\alias{genome2Genome}
\title{Map genomic positions in one genome to genomic positions in other genomes.}
\usage{
genome2Genome(ranges, aln, targetGenomes = NULL)
}
\arguments{
\item{ranges}{a \code{\linkS4class{GRanges}} or
\code{\linkS4class{GRangesList}} containing genomic
positions to map to the \code{targetGenomes}.}

\item{aln}{an \code{\linkS4class{annotatedAlignment}} instance.}

\item{targetGenomes}{a character vector of genome names.}
}
\value{
A \code{\linkS4class{GRangesList}} object.
}
\description{
Map genomic positions in one genome to genomic positions in other genomes.
}
\seealso{
\code{\link{alignment2Genome}}, \code{\link{genome2Alignment}},
\code{\link{sliceAlignment}}, \code{\link{findAnnotation}}.
}

