% Generated by roxygen2 (4.0.0): do not edit by hand
\name{importAlignment}
\alias{importAlignment}
\title{Import a genome alignment}
\usage{
importAlignment(alnpath, outfile = NULL)
}
\arguments{
\item{alnpath}{Path to the directory containing mercator segments,
a single file in \sQuote{mfa} (multi-fasta) or \sQuote{maf} (multiple
alignment file) format containing the genome aligment.}

\item{outfile}{if not \code{NULL} the alignment is writen to file in
multi-fasta format.}
}
\value{
A \code{\linkS4class{DNAStringSet}} instance.
}
\description{
Import a genome alignment
}
\details{
If the alignment is provided as a multi-fasta file the each header
must contain a map of the orthologous segments in the following
format:

GenomeID GenomeLength ChromosomeID:start:end:strand ... ChromosomeID:start:end:strand

where strand is encodes as + or - and GenomeLength is the length of the
genomes in the alignment which can be shorter than the length of the
original genome.
}
\seealso{
\code{\link{mercator}}.
}

