% Generated by roxygen2 (4.0.0): do not edit by hand
\name{install_genoslider_dependencies}
\alias{install_genoslider_dependencies}
\title{Install external dependencies for genoslideR}
\usage{
install_genoslider_dependencies(sudo = TRUE)
}
\arguments{
\item{sudo}{if \code{TRUE} install with sudo}
}
\description{
Works only on Ubuntu (and possibly Debian)
}

