\name{map2aln}
\alias{map2aln}
\title{INTERNAL: Map a genomic range to the alignment}
\usage{
  map2aln(ranges, aln)
}
\arguments{
  \item{range}{A \code{\linkS4class{Granges}} object.}

  \item{aln}{A \code{\linkS4class{AnnotatedAlignment}}
  object.}
}
\description{
  INTERNAL: Map a genomic range to the alignment
}
\keyword{internal}

