\name{map2aln}
\alias{map2aln}
\title{[INTERNAL] Map a genomic range to the alignment}
\usage{
  map2aln(ranges, gmap, amap, gaps)
}
\arguments{
  \item{ranges}{An IRanges instance}

  \item{gmap}{A Granges instance.}

  \item{amap}{A Granges instance.}

  \item{gaps}{A Granges instance.}
}
\value{
  A GrangesList instance.
}
\description{
  [INTERNAL] Map a genomic range to the alignment
}
\keyword{internal}

