% Generated by roxygen2 (4.0.0): do not edit by hand
\name{mercator}
\alias{mercator}
\title{Run mercator to build a homology map and create orthologous segments
that can be aligned using FSA}
\usage{
mercator(seq_files, anno_files = NULL, anno_type = "glimmer3",
  glimmeropts = list(o = 50, g = 110, t = 30), removeOverlappingCDS = TRUE,
  mask = TRUE, wd = NULL)
}
\arguments{
\item{seq_files}{Path to sequence files ('fasta' or 'genbank').}

\item{anno_files}{(Optional) Path to annotation files. If no
annotation files are provided an \emph{ab initio} annotation
using \code{\link{glimmer3}} is performed.}

\item{anno_type}{Type of annotation ('glimmer3', 'genbank', 'gff',
'ptt', or 'ftable').}

\item{glimmeropts}{Options passed to \code{\link{glimmer3}} if an
\emph{ab initio} annotation is performed.}

\item{removeOverlappingCDS}{Exclude overlapping CDS for orthology mapping.}

\item{mask}{Softmask sequence before aligning using
\code{\link{maskSequence}}.}

\item{wd}{Working directory. Defaults to the parent directory of
the provided sequence files.}
}
\description{
Run mercator to build a homology map and create orthologous segments
that can be aligned using FSA
}
\details{
\code{mercator} will create a hidden directory \sQuote{.mercator}
in the parent directory of the submitted sequence files and place
all results of the intermediate computational steps in this directory.

The results of a \code{mercator} run are stored in a directory
\sQuote{segments} in the parent directory of the submitted sequence files.
The \sQuote{segments} directory should contain numbered subdirectories
containing orthologous genomic segments in multi fasta format, a file
named \sQuote{map}, and a file named \sQuote{genomes}.

\code{mercator} segements can be aligned using the
\code{\link{alignSegments}} function, merged into a single multi-fasta
alignment file and imported into R as a \code{\linkS4class{DNAStringSet}}
using \code{\link{importAlignment}} or imported to R as an
\code{\linkS4class{annotatedAlignment}} using \code{\link{annotatedAlignment}}.
}
\seealso{
\code{\link{alignSegments}}, \code{\link{importAlignment}},
\code{\link{annotatedAlignment}}.
}

