% Generated by roxygen2 (4.0.0): do not edit by hand
\name{sliceAlignment}
\alias{sliceAlignment}
\title{Slice genomic ranges from an alignment.}
\usage{
sliceAlignment(ranges, aln, targetGenomes = NULL)
}
\arguments{
\item{ranges}{a \code{\linkS4class{GRanges}} or
\code{\linkS4class{GRangesList}} containing genomic positions.}

\item{aln}{an \code{\linkS4class{annotatedAlignment}} instance.}

\item{targetGenomes}{a character vector of genome names.}
}
\value{
A \code{\linkS4class{DNAStringSetList}} object.
}
\description{
The ranges passed to \code{sliceAlignment} are converted
from genomic ranges to alignment ranges. The alignment ranges
are returned as metadata with the resulting
\code{\linkS4class{DNAStringSetList}} object.
}
\seealso{
\code{\link{alignment2Genome}}, \code{\link{genome2Alignment}},
\code{\link{genome2Genome}}, \code{\link{findAnnotation}}.
}

